refactor: コード重複解消・責務分離・N+1クエリ最適化
- convex/helpers.ts と position/mutations.ts の重複関数を一本化
- shop/mutations.ts からスタッフ操作を staff/mutations.ts に分離
- staffSkill/queries.ts の listByShop でポジション一括取得に最適化
- PositionManager/PositionEditor の共通部品を抽出（SortablePositionItem, PositionAddForm, validatePositionName）
